+++ breaking_brain_api/urls.py
@@ -38,6 +38,7 @@ api_urlpatterns = [
     path('auth/', include('authentication.urls')),
     path('quizzes/', include('quizzes.urls')),
     path('users/', include('users.urls')),
+    path('search/', include('search.urls')),
 ]
 
 urlpatterns = [